Account for padding_idx in phone symbol list
This fixes an issue where <sil> was assigned to the 0th index in the
symbol list, which is also the padding_idx of the input symbol embedding
table. This meant that leading and trailing silences were effectively
ignored by the model, producing hard cuts at either end of synthesized
utterances.