Fix occasional mismatches between TextGrid durations and mels
This seems to affect utterances with sample lengths at integer multiples
of hop_length, where mel spectrograms end up with an extra frame
compared to frame durations calculated from TextGrid alignments. Seems
like this could be due to centering of sample windows during STFT, but
I'm not sure.

For some corpora, this doesn't seem to be an issue at all (e.g.
M-AILABS), while for others it shows up with some frequency (e.g.
GlobalPhone). There's some difference in the precision of WAV durations
reported by `soxi` between those two, which seems like it would line up
with the integer multiple thing, but again I don't know the reason for
that.

This seems not to be related to the occasional 2-frame mismatches for
durations calculated from run-length encoded HuBERT unit sequences,
since that issue arises on M-AILABS data which doesn't trigger this.