Rewrite TextProcessing
This restores a sensible option for processing raw text to train with
character inputs, rather than having to treat them as phone-like inputs
with appropriate preprocessing. This approach is simpler than the
original, with no support for ARPABET representation mixing (which
actually was all-or-nothing and doubled up as phone support in the
original implementation).

There are two ways to run this:
  - With MAS, passing raw text which will be straightforwardly expanded
    to character sequences, including spaces between words (which are
    lost when processing character-level forced alignments as phone-type
    inputs)
  - With `--symbol-set english_basic_sil`, which will handle silence
    phones from forced alignments as single tokens while still expanding
    other words to character sequences

TODO: The second option above probably still only works with MAS, since
it preserves spaces between words which will not be represented in
character-level forced alignments, so having no associated duration
targets and leading to mismatched sequence lengths between input symbols
and durations. So, to use character-level forced alignments probably
still requires treating text as phone-type inputs through preprocessing
to add spaces between characters.