Restore unbatched CTC alignment loss
This avoids an apparent problem with the batched CTC loss that I ran
into when trying to train on VCTK with MAS. For some batches, the CTC
backward pass would return nan. This seems to be triggered by the
increased proportion of shorter utterances in VCTK compared to other
corpora previously tested (around 45% of utterances are shorter than 4s;
training only with utterances > 3.5s long shows no problems), but I
couldn't pin it down completely.

This issue also never occurs for batch size 1, so feeding utterances to
the CTC loss one at a time also avoids the problem. Note this change is
restoring the original implementation from RAD-TTS.